Source code for the DOS version of the FFT(MAX) system.
The files are from one of the earliest DOS source archives I could find. Note
that the timestamp info and version number (1.35) of these files is more recent
than the publication date of the textbook, and the version of the binaries
(1.31) packaged with it.

In addition to source files, the archive also contained some build artifacts
(intermediate .trc bytecode and debug files.) I've kept them because they may
give some clues as to, or verification of, the source files that produced them.
In some cases, the .trc file is all we have.

The files in Lib/ should be included with the source files shipped with a
True BASIC install; they're included here in case they were modified or these
versions of the files aren't available any more.

Files are encoded in DOS codepage 437.