Start aligning the DOS sources towards the Mac FFT sources.
This is done to make it easier to discern the functional differences between the
Mac and DOS codebases later - as opposed to changes caused by, for example,
splitting files or moving functions around.

As a first step, this check-in clones several large files which will have their
functions divided into several smaller, more manageable ones.