Moved subroutines between files.
Up to this point, there have been no functional changes to the source code, only
re-ordering. So it should be possible - perhaps with a bit of jiggling of
library imports and function calls - to compile this version of the source tree
and get the same executable as the last full DOS source commit.