Source code for the Macintosh version of the FFT system.
Judging by the version information in LIBINFO.TRU, this is one of the earliest
FFT source archives I could unearth. The timestamp of the archive and the files
within, however, is 2000-02-04 12:41 - but this would be when I originally
transferred the files off Dr. Morrison's Macintosh. The original timestamps
weren't preserved in the move.